requirements.txt
pandas
requests
beautifulsoup4
apache-airflow